Reorganize folder structure to put classes in folders for each Clean Architecture layer. Change MetricValues to store ArrayList for its two attributes instead of List to enable getting the last item (which is not supported by List).
