+++ src/main/java/ui/StockDataView.java
@@ -1,11 +1,5 @@
-package main.java.view;
+package main.java.ui;
 
-import java.awt.*;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.swing.*;
